feat: responsive exhibitions landing page
