+++ components/ExhibitionsComponents/AllExhibitions/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 
 import FeatureHeader from "../../shared/FeatureHeader";
-import Exhibitions from "./components/Exhibitions";
+import ExhibitionsList from "./components/ExhibitionsList";
 import mockExhibitions from "./mockExhibitions";
 
 import { TITLE, DESCRIPTION } from "../../../constants/exhibitions.js";
@@ -9,7 +9,7 @@ import { TITLE, DESCRIPTION } from "../../../constants/exhibitions.js";
 const AllSets = ({ route }) =>
   <div>
     <FeatureHeader title={TITLE} description={DESCRIPTION} />
-    <Exhibitions exhibitions={mockExhibitions} route={route} />
+    <ExhibitionsList exhibitions={mockExhibitions} route={route} />
   </div>;
 
 export default AllSets;
